nj,fngv

+++ S32K144_ST7789_Graphics_Test/S&M_Driv/inc/ST7789_low_level.h
@@ -111,7 +111,7 @@ void ST7789_Draw_Pixel_4px( uint16_t x, uint16_t y, uint16_t color);
 void ST7789_DrawLine( uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 void ST7789_DrawRectangle(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2, uint16_t color);
 void ST7789_DrawCircle(uint16_t x0, uint16_t y0, uint8_t r, uint16_t color);
-void ST7789_DrawImage(uint16_t x, uint16_t y, uint16_t w, uint16_t h, const uint16_t *data);
+void ST7789_DrawImage(uint16_t x, uint16_t y, uint16_t w, uint16_t h, const uint8_t *data);
 void ST7789_InvertColors(uint8_t invert);
 void ST7789_WriteChar(uint16_t x, uint16_t y, char ch, FontDef font, uint16_t color, uint16_t bgcolor);
 void ST7789_WriteString(uint16_t x, uint16_t y, const char *str, FontDef font, uint16_t color, uint16_t bgcolor);
